Implement day 08
It is a O(n*sqrt(n)) solution (we loop over each tree, and for each tree we
consider every other tree in the row/column.

For part 1 I initially had a O(n) solution, but I got rid of it as it
didn't generalize to part 2.

I had a preloop in each of the 4 directions which would store the
tallest tree so far. Then the 2nd loop could refer to those pre-calcs.