Solve 2024 day 24 part 2
I solved it, but using the code as a helper, rather than fully solving
programatically. In notes I worked out that for z02 and above the
following applies:

zn = (xn xor yn) xor carry n-1
and carry n = (xn AND yn) OR (carry n -1 AND (xn XOR yn)).

note, that although the problem didn't produce it, the XOR could be an
OR.

That is to say:
carries are recursive, so when fully expanded carries go all the way
down to x0 and y0, so they get very long and it's not constructive to
look at them expanded
and
carries are the only thing produced by the OR operator.

Printing out the equations with this substition was enough to identify
every single swap that needed making, although the other half of the
swap required looking further in.

I didn't program this beyond printing out the equations and making the
swaps.